+++ lib/squid.rb
@@ -1,0 +1,5 @@
+require 'prawn'
+require 'squid/interface'
+
+module Squid
+end