+++ manual/contents.rb
@@ -1,0 +1,7 @@
+require 'squid'
+require 'prawn/manual_builder'
+Prawn::ManualBuilder.manual_dir = File.dirname(__FILE__)
+Prawn::Font::AFM.hide_m17n_warning = true
+Prawn::ManualBuilder::Example.generate 'manual.pdf', skip_page_creation: true do
+  load_package 'squid'
+end