+++ spec/spec_helper.rb
@@ -7,14 +7,12 @@ SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter[
 ]
 SimpleCov.start
 
-# ----------
-
-require 'pdf/inspector'
-require 'prawn/chart'
-
 Dir['./spec/support/**/*.rb'].each {|f| require f}
 
 RSpec.configure do |config|
   config.order = 'random'
   config.run_all_when_everything_filtered = false
 end
+
+require 'pdf/inspector'
+require 'squid'